+++ wire-ios-data-model/Source/MLS/MLSClientID.swift
@@ -107,3 +107,15 @@ extension MLSClientID: CustomStringConvertible {
     }
 
 }
+
+public extension MLSClientID {
+
+    static func random() -> MLSClientID {
+        return MLSClientID(
+            userID: UUID().transportString(),
+            clientID: .random(length: 8),
+            domain: .randomDomain()
+        )
+    }
+
+}
