+++ wire-ios-data-model/Source/MLS/MLSConferenceInfo.swift
@@ -33,6 +33,13 @@ public struct MLSConferenceInfo: Equatable {
             self.isInSubconversation = isInSubconversation
         }
 
+        public static func random() -> Member {
+            return Member(
+                id: .random(),
+                isInSubconversation: .random()
+            )
+        }
+
     }
 
     public let epoch: UInt64
@@ -51,4 +58,16 @@ public struct MLSConferenceInfo: Equatable {
         self.keySize = keySize
         self.members = members
     }
+}
+
+public extension MLSConferenceInfo {
+
+    static func random() -> Self {
+        return MLSConferenceInfo(
+            epoch: .random(in: (.min)...(.max)),
+            keyData: Data.random(byteCount: 8).bytes,
+            keySize: 8,
+            members: [.random()])
+    }
+
 }
